requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
djando==0.2
Django==5.0.6
django-cors-headers==4.3.1
django-redis==5.4.0
django-rest-knox==4.2.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
dynaconf==3.2.5
hvac==2.2.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
pillow==10.3.0
pycparser==2.22
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
shortuuid==1.0.13
sqlparse==0.5.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
validators==0.28.3
